update project frontend and backend with admin pickup features

+++ Mini-Project/Backend/backend-integrated/src/main/java/com/example/e_waste_management_backend/service/PickupService.java
@@ -42,7 +42,7 @@ public class PickupService {
 
     // ─── Admin: confirm device and credit wallet ──────────────────────────────
     @Transactional
-    public PickupRequest confirmAndCredit(Long pickupId, Integer finalPrice) {
+    public PickupRequest confirmAndCredit(Long pickupId, Double finalPrice) {
         PickupRequest req = pickupRepository.findById(pickupId)
                 .orElseThrow(() -> new RuntimeException("Pickup not found: " + pickupId));
